RHM-47 Create two different scripts: rhdzmota and rhdzmota.ext
